src/config.py
import os
from pathlib import Path

from dotenv import load_dotenv

load_dotenv()

root = os.getenv("ROOT")
if root is None:
    raise ValueError(
        "ROOT environment variable is not set in .env file. Use dot-env-template file to create .env file."
    )

ROOT_DIR = Path(root)

DATA_DIR = ROOT_DIR / "data"
RAW_DATA_DIR = DATA_DIR / "raw"
INTERIM_DATA_DIR = DATA_DIR / "interim"
PROCESSED_DATA_DIR = DATA_DIR / "processed"
EXTERNAL_DATA_DIR = DATA_DIR / "external"

FIGURES_DIR = ROOT_DIR / "reports" / "figures"
